Abstraction, abstraction - added Main,the code has been moved to the package,edit attack logic, constructor all parameters
