+++ src/main/java/school/faang/BJS2_43856/Warrior.java
@@ -1,0 +1,19 @@
+package school.faang.BJS2_43856;
+
+import lombok.NonNull;
+
+public class Warrior extends Character {
+    public Warrior(String name) {
+        super(name, 10, 5, 3);
+    }
+
+    @Override
+    public void attack(@NonNull Character character) {
+        if(character.getHealth() - this.getPower() > 0){
+            character.setHealth(character.getHealth() - this.getPower());
+        }
+        else{
+            character.setHealth(0);
+        }
+    }
+}
