--- src/main/java/school/faang/Warrior.java
@@ -1,17 +1,0 @@
-package school.faang;
-
-import lombok.NonNull;
-
-public class Warrior extends Character {
-    public Warrior(String name) {
-        super(name);
-        this.setPower(10);
-        this.setDexterity(5);
-        this.setIntelligence(3);
-    }
-
-    @Override
-    public void attack(@NonNull Character character) {
-        this.setHealth(this.getHealth() - character.getPower());
-    }
-}
